feat: reorganized frontend api
