+++ Frontend/src/api/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
 export * from './endpoints'
+export * from './client'
