+++ Frontend/src/api/types/common.ts
@@ -1,0 +1,5 @@
+// components['schemas']['TagRequest']
+export type TagRequest = {
+    tagName: string;
+    isSkill: boolean;
+}
