+++ Frontend/src/api/types/users.ts
@@ -1,0 +1,34 @@
+import { components } from '../schema';
+
+// components['schemas']['UserResponse']
+export type User = {
+    id: string;
+    name: string;
+    email: string;
+    interestTags: string[];
+    skillTags: string[];
+}
+
+// components['schemas']['UserRequest']
+export type UserRequest = {
+    clerkId: string;
+    name: string;
+    email: string;
+}
+
+// components['schemas']['UserPatchRequest'];
+export type UserPatchRequest = {
+    clerkId: string;
+    skillTags?: string[];
+    interestTags?: string[];
+}
+
+export function mapUserResponseToUser(userResponse: components['schemas']['UserResponse']) {
+    return {
+        id: userResponse.id!,
+        name: userResponse.name!,
+        email: userResponse.email!,
+        interestTags: userResponse.interestTags!,
+        skillTags: userResponse.skillTags!
+    } as User;
+}
